add script for running 'go test --cover'
