+++ gpMgmt/go-utils/src/gp_upgrade/Makefile
@@ -15,9 +15,11 @@ dependencies :
 		go get github.com/cppforlife/go-semi-semantic/version
 		go get github.com/onsi/ginkgo/ginkgo
 		go get golang.org/x/tools/cmd/goimports
-		go get github.com/maxbrunsfeld/counterfeiter
 		go get github.com/onsi/gomega
 		go get github.com/jessevdk/go-flags
+# Counterfeiter is not a proper dependency of the app. It is only used occasionally to generate a test class that
+# is then checked in.  At the time of that generation, it can be added back to run the dependency list, temporarily.
+#		go get github.com/maxbrunsfeld/counterfeiter
 
 format : dependencies
 		goimports -w .
@@ -39,6 +41,10 @@ push : format
 
 build : dependencies
 		go build -ldflags "-X gp_upgrade/commands.GpdbVersion=$(GPDB_VERSION)" -o $(GOPATH)/bin/$(MODULE_NAME)
+
+coverage: dependencies format sshd_build build 
+		./scripts/run_coverage.sh
+
 
 linux :
 		GOOS=$@ GOARCH=$(ARCH) go build -ldflags "-X gp_upgrade/commands.GpdbVersion=$(GPDB_VERSION)" -o $(GOPATH)/bin/$(MODULE_NAME).$@
